move updater to bridge page
