+++ src/index.tsx
@@ -25,7 +25,6 @@ import App from './pages/App'
 import * as serviceWorkerRegistration from './serviceWorkerRegistration'
 import store from './state'
 import ApplicationUpdater from './state/application/updater'
-import BridgeUpdater from './state/bridge/updater'
 import CustomizeDexesUpdater from './state/customizeDexes/updater'
 import ListsUpdater from './state/lists/updater'
 import MulticallUpdater from './state/multicall/updater'
@@ -69,7 +68,6 @@ function Updaters() {
   return (
     <>
       <ListsUpdater />
-      <BridgeUpdater />
       <UserUpdater />
       <ApplicationUpdater />
       <TransactionUpdater />
