+++ src/utils/tokenInfo.ts
@@ -1,9 +1,8 @@
-import { ChainId, Currency } from '@kyberswap/ks-sdk-core'
+import { ChainId, Currency, WETH } from '@kyberswap/ks-sdk-core'
 import { getAddress } from 'ethers/lib/utils'
 
 import { NETWORKS_INFO } from 'constants/networks'
 import { MAP_TOKEN_HAS_MULTI_BY_NETWORK, WHITE_LIST_TOKEN_INFO_PAIR } from 'constants/tokenLists/token-info'
-import { nativeOnChain } from 'constants/tokens'
 import { WrappedTokenInfo } from 'state/lists/wrappedTokenInfo'
 
 /**
@@ -57,7 +56,7 @@ export const isTokenNative = (currency: Currency | WrappedTokenInfo | undefined,
   if (currency?.isNative) return true
   // case multichain token
   return chainId
-    ? nativeOnChain(chainId).wrapped.address === currency?.address &&
+    ? WETH[chainId]?.address === currency?.address &&
         currency instanceof WrappedTokenInfo &&
         currency.multichainInfo?.tokenType === 'NATIVE'
     : false
